+++ src/HumdrumToken.cpp
@@ -1835,7 +1835,7 @@ int HumdrumToken::addLinkedParameter(HTp token) {
 	}
 
 	m_linkedParameters.push_back(token);
-	return m_linkedParameters.size() - 1;
+	return (int)m_linkedParameters.size() - 1;
 }
